9_bot_in_docker/requirements.txt
astroid==2.9.0
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.4.0
pylint==2.12.2
python-dotenv==0.19.2
requests==2.26.0
toml==0.10.2
typed-ast==1.5.1
typing_extensions==4.0.1
urllib3==1.26.7
wrapt==1.13.3
